refactor: decouple action-management
