+++ src/action-management/types.d.ts
@@ -7,7 +7,13 @@ interface ActionManager {
 
 type Condition = boolean | error;
 
+interface Core {
+  setFailed(msg: string): void;
+  warning(msg: string): void;
+}
+
 export type {
   ActionManager,
   Condition,
+  Core,
 };
